Fix for double-saving bug. Very confusing!
Here's the explanation:

I was handling the save with (1) the onclick event of the save button AND (2) the return-key keyup event for the row. If you tabbed to "save" and hit enter, I hypothesized that both the click AND the enter event were saving the same row. What was actually happening was:

1. onClick is called for the new row
 2. save is requested
 3. a second new add row is immediately created and added to the DOM
 4. focus is given to the new add row
 5. onClick event finishes for the previous row
6. keyup event is called, but now the focus is in the body textbox in the new add row
 7. save is requested on new row because enter key was pressed; naturally this row is empty
 8. a third new row is added and focus is given there.
 
This seems like a browser failing because the WRONG OBJECT gets the enter-keypress event. In this case, event.element() is NOT the element you'd expect, since you actually pressed enter on the BUTTON. 





git-svn-id: svn+ssh://taylor-hughes.com/svnroot/dnzo/trunk@69 c02d1e6f-6a35-45f2-ab14-3b6f79a691ff