Add "transparent" timezone setting.
When the user logs in and loads his task list, the JS now checks his timezone setting and submits a different setting if the setting is wrong. This should all happen before a timezone is actually needed for anything and will Just Work(tm) with DST, different computers, etc., as long as the user's machine has the proper time zone setting itself.






git-svn-id: svn+ssh://taylor-hughes.com/svnroot/dnzo/trunk@141 c02d1e6f-6a35-45f2-ab14-3b6f79a691ff