Add notion of sortability and remove start and end effects.
Not all lists will be re-sortable, since if you're sorted by a column that doesn't make sense. So, if we're not sorted by anything (default sort), add a class to the table to let the JS know it's re-sortable, and use that value to (a) set cursor: move and (b) decide if we should wire our sorting stuff.




git-svn-id: svn+ssh://taylor-hughes.com/svnroot/dnzo/trunk@195 c02d1e6f-6a35-45f2-ab14-3b6f79a691ff